# Improve performance of looking up signatures.