Turns out iterating byte-wise through a huge binary in python is very
slow.

Singular data point for a call to `get_addr_from_sig` for a signature on a
40 MB binary with ~125k functions:

runtime in [s]     |     old |  new
----------------|---------:|------------:
sig in binary      |  44.74  | 0.92
sig not in binary  | 167.23  | 4.42